my_project/backend/requirements.txt
Django>=4.2,<5.0
psycopg2>=2.9,<3.0
djangorestframework>=3.14,<4.0
django-cors-headers>=3.10,<4.0  # Mettez à jour cette ligne
